new issue form service refactor (#224242)
* fix merge conflicts and refactor on issue form

* large rehaul, removing more stuff, merging into main base issue reporter service now

* more cleanup

* remove unused window configuration

* some cleanup

* more cleanup
